Reorganize compiler sources into src/compiler
