+++ src/compiler/compdiag.c
