+++ src/compiler/emitbc.c
